#865 reformat interface submodels
