+++ docs/source/models/submodels/interface/diffusion_limited.rst
@@ -1,0 +1,5 @@
+Diffusion-limited
+=================
+
+.. autoclass:: pybamm.interface.DiffusionLimited
+    :members:
