--- docs/source/models/submodels/interface/diffusion_limited/base_diffusion_limited.rst
@@ -1,5 +1,0 @@
-Base Model
-==========
-
-.. autoclass:: pybamm.interface.diffusion_limited.BaseModel
-    :members:
