--- docs/source/models/submodels/interface/diffusion_limited/full_diffusion_limited.rst
@@ -1,5 +1,0 @@
-Full Model
-==========
-
-.. autoclass:: pybamm.interface.diffusion_limited.FullDiffusionLimited
-    :members:
