--- docs/source/models/submodels/interface/diffusion_limited/index.rst
@@ -1,9 +1,0 @@
-Diffusion-limited Kinetics
-==========================
-
-.. toctree::
-  :maxdepth: 1
-
-  base_diffusion_limited
-  full_diffusion_limited
-  leading_diffusion_limited
