--- docs/source/models/submodels/interface/diffusion_limited/leading_diffusion_limited.rst
@@ -1,5 +1,0 @@
-Leading-order Model
-===================
-
-.. autoclass:: pybamm.interface.diffusion_limited.LeadingOrderDiffusionLimited
-    :members:
