+++ docs/source/models/submodels/interface/first_order_kinetics/first_order_kinetics.rst
@@ -1,0 +1,5 @@
+First-order Kinetics
+====================
+
+.. autoclass:: pybamm.interface.first_order_kinetics.FirstOrderKinetics
+    :members:
