+++ docs/source/models/submodels/interface/first_order_kinetics/index.rst
@@ -1,0 +1,8 @@
+First-order Kinetics
+====================
+
+.. toctree::
+  :maxdepth: 1
+
+  inverse_first_order_kinetics
+  first_order_kinetics
