+++ docs/source/models/submodels/interface/inverse_kinetics/index.rst
@@ -4,6 +4,4 @@ Inverse Interface Kinetics
 .. toctree::
   :maxdepth: 1
 
-  base_inverse_first_order_kinetics
-  base_inverse_kinetics
   inverse_butler_volmer
