--- docs/source/models/submodels/interface/kinetics/base_first_order_kinetics.rst
@@ -1,5 +1,0 @@
-Base First-order Kinetics
-=========================
-
-.. autoclass:: pybamm.interface.kinetics.BaseFirstOrderKinetics
-    :members:
