+++ docs/source/models/submodels/interface/kinetics/butler_volmer.rst
@@ -3,6 +3,3 @@ Butler-Volmer
 
 .. autoclass:: pybamm.interface.kinetics.ButlerVolmer
     :members:
-
-.. autoclass:: pybamm.interface.kinetics.FirstOrderButlerVolmer
-    :members:
