+++ docs/source/models/submodels/interface/kinetics/index.rst
@@ -5,7 +5,6 @@ Interface Kinetics
   :maxdepth: 1
 
   base_kinetics
-  base_first_order_kinetics
   butler_volmer
   no_reaction
   tafel
