+++ docs/source/models/submodels/interface/kinetics/tafel.rst
@@ -3,9 +3,6 @@ Tafel
 
 .. autoclass:: pybamm.interface.kinetics.ForwardTafel
     :members:
-    
-.. autoclass:: pybamm.interface.kinetics.FirstOrderForwardTafel
-    :members:
 
 .. autoclass:: pybamm.interface.kinetics.BackwardTafel
     :members:
