--- docs/source/models/submodels/interface/lead_acid.rst
@@ -1,14 +1,0 @@
-Lead Acid
-=========
-
-.. autoclass:: pybamm.interface.lead_acid.ButlerVolmer
-    :members:
-
-.. autoclass:: pybamm.interface.lead_acid.InverseButlerVolmer
-    :members:
-
-.. autoclass:: pybamm.interface.lead_acid.FirstOrderButlerVolmer
-    :members:
-
-.. autoclass:: pybamm.interface.lead_acid.InverseFirstOrderKinetics
-    :members:
