--- docs/source/models/submodels/interface/lithium_ion.rst
@@ -1,8 +1,0 @@
-Lithium-Ion
-===========
-
-.. autoclass:: pybamm.interface.lithium_ion.ButlerVolmer
-    :members:
-    
-.. autoclass:: pybamm.interface.lithium_ion.InverseButlerVolmer
-    :members:
