+++ pybamm/models/full_battery_models/lead_acid/full.py
@@ -61,11 +61,11 @@ class Full(BaseModel):
             self.submodels["convection"] = pybamm.convection.Full(self.param)
 
     def set_interfacial_submodel(self):
-        self.submodels["negative interface"] = pybamm.interface.lead_acid.ButlerVolmer(
-            self.param, "Negative"
+        self.submodels["negative interface"] = pybamm.interface.ButlerVolmer(
+            self.param, "Negative", "lead-acid main"
         )
-        self.submodels["positive interface"] = pybamm.interface.lead_acid.ButlerVolmer(
-            self.param, "Positive"
+        self.submodels["positive interface"] = pybamm.interface.ButlerVolmer(
+            self.param, "Positive", "lead-acid main"
         )
 
     def set_solid_submodel(self):
@@ -108,22 +108,22 @@ class Full(BaseModel):
             self.submodels["oxygen diffusion"] = pybamm.oxygen_diffusion.Full(
                 self.param, self.reactions
             )
-            self.submodels[
-                "positive oxygen interface"
-            ] = pybamm.interface.lead_acid_oxygen.ForwardTafel(self.param, "Positive")
+            self.submodels["positive oxygen interface"] = pybamm.interface.ForwardTafel(
+                self.param, "Positive", "lead-acid oxygen"
+            )
             self.submodels[
                 "negative oxygen interface"
-            ] = pybamm.interface.lead_acid_oxygen.FullDiffusionLimited(
-                self.param, "Negative"
+            ] = pybamm.interface.DiffusionLimited(
+                self.param, "Negative", "lead-acid oxygen", order="full"
             )
         else:
             self.submodels["oxygen diffusion"] = pybamm.oxygen_diffusion.NoOxygen(
                 self.param
             )
-            self.submodels[
-                "positive oxygen interface"
-            ] = pybamm.interface.lead_acid_oxygen.NoReaction(self.param, "Positive")
-            self.submodels[
-                "negative oxygen interface"
-            ] = pybamm.interface.lead_acid_oxygen.NoReaction(self.param, "Negative")
+            self.submodels["positive oxygen interface"] = pybamm.interface.NoReaction(
+                self.param, "Positive", "lead-acid oxygen"
+            )
+            self.submodels["negative oxygen interface"] = pybamm.interface.NoReaction(
+                self.param, "Negative", "lead-acid oxygen"
+            )
