+++ pybamm/models/full_battery_models/lead_acid/loqs.py
@@ -122,21 +122,31 @@ class LOQS(BaseModel):
         if self.options["surface form"] is False:
             self.submodels[
                 "leading-order negative interface"
-            ] = pybamm.interface.lead_acid.InverseButlerVolmer(self.param, "Negative")
+            ] = pybamm.interface.InverseButlerVolmer(
+                self.param, "Negative", "lead-acid main"
+            )
             self.submodels[
                 "leading-order positive interface"
-            ] = pybamm.interface.lead_acid.InverseButlerVolmer(self.param, "Positive")
+            ] = pybamm.interface.InverseButlerVolmer(
+                self.param, "Positive", "lead-acid main"
+            )
         else:
             self.submodels[
                 "leading-order negative interface"
-            ] = pybamm.interface.lead_acid.ButlerVolmer(self.param, "Negative")
+            ] = pybamm.interface.ButlerVolmer(self.param, "Negative", "lead-acid main")
 
             self.submodels[
                 "leading-order positive interface"
-            ] = pybamm.interface.lead_acid.ButlerVolmer(self.param, "Positive")
+            ] = pybamm.interface.ButlerVolmer(self.param, "Positive", "lead-acid main")
+        # always use forward Butler-Volmer for the reaction submodel to be passed to the
+        # higher order model
         self.reaction_submodels = {
-            "Negative": [self.submodels["leading-order negative interface"]],
-            "Positive": [self.submodels["leading-order positive interface"]],
+            "Negative": [
+                pybamm.interface.ButlerVolmer(self.param, "Negative", "lead-acid main")
+            ],
+            "Positive": [
+                pybamm.interface.ButlerVolmer(self.param, "Positive", "lead-acid main")
+            ],
         }
 
     def set_negative_electrode_submodel(self):
@@ -186,22 +196,24 @@ class LOQS(BaseModel):
             ] = pybamm.oxygen_diffusion.LeadingOrder(self.param, self.reactions)
             self.submodels[
                 "leading-order positive oxygen interface"
-            ] = pybamm.interface.lead_acid_oxygen.ForwardTafel(self.param, "Positive")
+            ] = pybamm.interface.ForwardTafel(
+                self.param, "Positive", "lead-acid oxygen"
+            )
             self.submodels[
                 "leading-order negative oxygen interface"
-            ] = pybamm.interface.lead_acid_oxygen.LeadingOrderDiffusionLimited(
-                self.param, "Negative"
+            ] = pybamm.interface.DiffusionLimited(
+                self.param, "Negative", "lead-acid oxygen", order="leading"
             )
         else:
             self.submodels[
                 "leading-order oxygen diffusion"
             ] = pybamm.oxygen_diffusion.NoOxygen(self.param)
             self.submodels[
-                "leading-order positive oxygen interface"
-            ] = pybamm.interface.lead_acid_oxygen.NoReaction(self.param, "Positive")
-            self.submodels[
                 "leading-order negative oxygen interface"
-            ] = pybamm.interface.lead_acid_oxygen.NoReaction(self.param, "Negative")
+            ] = pybamm.interface.NoReaction(self.param, "Negative", "lead-acid oxygen")
+            self.submodels[
+                "leading-order positive oxygen interface"
+            ] = pybamm.interface.NoReaction(self.param, "Positive", "lead-acid oxygen")
         self.reaction_submodels["Negative"].append(
             self.submodels["leading-order negative oxygen interface"]
         )
