+++ pybamm/models/submodels/interface/__init__.py
@@ -1,4 +1,5 @@
 from .base_interface import BaseInterface
-from . import lead_acid
-from . import lead_acid_oxygen
-from . import lithium_ion
+from .kinetics import *
+from .inverse_kinetics import *
+from .first_order_kinetics import *
+from .diffusion_limited import DiffusionLimited
