--- pybamm/models/submodels/interface/diffusion_limited/__init__.py
@@ -1,3 +1,0 @@
-from .base_diffusion_limited import BaseModel
-from .leading_diffusion_limited import LeadingOrderDiffusionLimited
-from .full_diffusion_limited import FullDiffusionLimited
