+++ pybamm/models/submodels/interface/first_order_kinetics/__init__.py
@@ -1,0 +1,2 @@
+from .first_order_kinetics import FirstOrderKinetics
+from .inverse_first_order_kinetics import InverseFirstOrderKinetics
