+++ pybamm/models/submodels/interface/inverse_kinetics/__init__.py
@@ -1,3 +1,1 @@
-from .base_inverse_kinetics import BaseInverseKinetics
-from .base_inverse_first_order_kinetics import BaseInverseFirstOrderKinetics
 from .inverse_butler_volmer import InverseButlerVolmer
