+++ pybamm/models/submodels/interface/kinetics/__init__.py
@@ -1,5 +1,4 @@
-from .base_kinetics import BaseModel
-from .base_first_order_kinetics import BaseFirstOrderKinetics
-from .butler_volmer import ButlerVolmer, FirstOrderButlerVolmer
-from .tafel import ForwardTafel, BackwardTafel, FirstOrderForwardTafel
+from .base_kinetics import BaseKinetics
+from .butler_volmer import ButlerVolmer
+from .tafel import ForwardTafel, BackwardTafel
 from .no_reaction import NoReaction
