+++ pybamm/models/submodels/interface/kinetics/butler_volmer.py
@@ -3,11 +3,10 @@
 #
 
 import pybamm
-from .base_kinetics import BaseModel
-from .base_first_order_kinetics import BaseFirstOrderKinetics
+from .base_kinetics import BaseKinetics
 
 
-class ButlerVolmer(BaseModel):
+class ButlerVolmer(BaseKinetics):
     """
     Base submodel which implements the forward Butler-Volmer equation:
 
@@ -20,20 +19,21 @@ class ButlerVolmer(BaseModel):
         model parameters
     domain : str
         The domain to implement the model, either: 'Negative' or 'Positive'.
+    reaction : str
+        The name of the reaction being implemented
 
-
-    **Extends:** :class:`pybamm.interface.kinetics.BaseModel`
+    **Extends:** :class:`pybamm.interface.kinetics.BaseKinetics`
     """
 
-    def __init__(self, param, domain):
-        super().__init__(param, domain)
+    def __init__(self, param, domain, reaction):
+        super().__init__(param, domain, reaction)
 
     def _get_kinetics(self, j0, ne, eta_r, T):
         prefactor = ne / (2 * (1 + self.param.Theta * T))
         return 2 * j0 * pybamm.sinh(prefactor * eta_r)
 
     def _get_dj_dc(self, variables):
-        "See :meth:`pybamm.interface.kinetics.BaseModel._get_dj_dc`"
+        "See :meth:`pybamm.interface.kinetics.BaseKinetics._get_dj_dc`"
         c_e, delta_phi, j0, ne, ocp, T = self._get_interface_variables_for_first_order(
             variables
         )
@@ -44,15 +44,10 @@ class ButlerVolmer(BaseModel):
         )
 
     def _get_dj_ddeltaphi(self, variables):
-        "See :meth:`pybamm.interface.kinetics.BaseModel._get_dj_ddeltaphi`"
+        "See :meth:`pybamm.interface.kinetics.BaseKinetics._get_dj_ddeltaphi`"
         _, delta_phi, j0, ne, ocp, T = self._get_interface_variables_for_first_order(
             variables
         )
         eta_r = delta_phi - ocp
         prefactor = ne / (2 * (1 + self.param.Theta * T))
         return 2 * j0 * prefactor * pybamm.cosh(prefactor * eta_r)
-
-
-class FirstOrderButlerVolmer(ButlerVolmer, BaseFirstOrderKinetics):
-    def __init__(self, param, domain):
-        super().__init__(param, domain)
